fix(glyph): keep template source files within the length budget
The attribute-order change pushed template.rs past its base size and
tipped attributes.rs over the 350-line limit, failing the source-file
-lengths guard. Compact attribute_priority (matches! + tighter doc) so
attributes.rs is back under 350, and relocate its unit test into a new
attribute_priority_tests module so the already-over-limit template.rs
no longer grows. No behavior change.
